Improve API for render passes.
Note for Vulkan: I'm assuming that if Vulkan is ever implemented,
CeresGPU will do resource tracking an insert barriers as needed.

Later, CeresGPU could have a manual syncronization API for those needing
more fine grained control. Hopefully this would come with a
syncronization validator to reduce the ammount of works-on-my-machine
bugs that come with Vulkan syncronization.