Add ISampler support for Metal.
NOTE: There's some inconsistency between Metal and OpenGL.
Metal will not sample textures at all if the sampler is not set.
OpenGL just defaults back to the texture object's settings.

We should add checks or a way for descriptor set instances
to fall back to a default sampler if one wasn't set.