Some initial stubs of some 'Verification' wrapper classes.
May or may do verification this way, however verifying that Streaming
buffers are being used correctly is difficult, since actual access of
the buffers is done on the GPU, making it hard for us to verify the
usage of the buffers from our Verification wrapper resources.

Maybe there is a way that relies on the user being more explicit
about the use of their buffers.